refactor(core-weave): extract request and candidate model types
- move core weave request contracts into requests.ts while preserving weave.ts type re-exports
- move reusable source locator, candidate, planning, and slice-name types into focused core weave modules
- update runtime weave type imports to use the focused model modules directly
- keep planner functions and generated RDF/Page behavior unchanged
- verify with fmt, lint, check, core weave tests, integration weave tests, and import-cycle audit
