+++ src/core/weave/slices.ts
@@ -1,0 +1,7 @@
+export type WeaveSlice =
+  | "firstKnopWeave"
+  | "firstPayloadWeave"
+  | "firstExtractedKnopWeave"
+  | "firstReferenceCatalogWeave"
+  | "pageDefinitionWeave"
+  | "secondPayloadWeave";
