+++ src/core/weave/source_models.ts
@@ -1,0 +1,4 @@
+export interface RepositorySourceFloatingLocator {
+  repositoryUrl: string;
+  repositoryPathFromRoot: string;
+}
