+++ src/runtime/weave/raw_source_panels.ts
@@ -5,9 +5,9 @@ import {
   toKnopPath,
 } from "../../core/designator_segments.ts";
 import { SFLO_NAMESPACE } from "../../core/rdf/namespaces.ts";
+import type { RepositorySourceFloatingLocator } from "../../core/weave/source_models.ts";
 import type {
   KnopArtifactLinkModel,
-  RepositorySourceFloatingLocator,
   ResourcePageRawSourcePanelModel,
 } from "../../core/weave/resource_page_models.ts";
 import {
